Fixed compilation warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@99031 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/pubseq_gateway/client/psg_client.cpp
@@ -424,7 +424,7 @@ CPSG_ReplyItem* CPSG_Reply::SImpl::CreateImpl(SPSG_Reply::SItem::TTS& item_ts, S
         _TROUBLE;
 
     } else if (type != CPSG_ReplyItem::eEndOfReply) {
-        if (stats) stats->IncCounter(SPSG_Stats::eReplyItemStatus, static_cast<size_t>(status));
+        if (stats) stats->IncCounter(SPSG_Stats::eReplyItemStatus, static_cast<unsigned>(status));
         return new CPSG_ReplyItem(type);
     }
